backend/requirements.txt
boto3==1.26.137
Flask==2.3.2
python-dotenv==1.0.0
flask-cors==4.0.0
PyJWT==2.7.0